third_party/blink/web_tests/wpt_internal/webgl/draw-element-image/nested-webgl-layoutsubtree-canvas.html

<!DOCTYPE html>
<title>texElementImage2D should throw if the canvas is nested inside another canvas</title>
<link rel="help" href="https://github.com/WICG/html-in-canvas">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/wpt_internal/resources/canvas-draw-element/waitForCanvasPaint.js"></script>

<canvas id="webgl_canvas" layoutsubtree="true" width="100" height="100">
  <canvas id="nested_webgl_canvas" layoutsubtree="true" width="100" height="100">
    <div id="inner_div" style="width: 100px; height: 100px;"></div>
  </canvas>
</canvas>
<script>
promise_test(async (t) => {
  await waitForCanvasPaint(webgl_canvas);
  const gl = nested_webgl_canvas.getContext('webgl2', {preserveDrawingBuffer:true});
  const tex = gl.createTexture();
  gl.bindTexture(gl.TEXTURE_2D, tex);
  assert_throws_dom("NotSupportedError", () => {
    gl.texElementImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, inner_div);
  });
}, "texElementImage2D should throw if called on a nested webgl canvas");
</script>
